Finished the base game play. I will need to add in challege stages and maybe find a way to make the aliens shoot missiles more aggressively. Maybe something like randomly pic a y level and when they are at that y level they shoot and as the levels progress, they get more y levels to shoot from.
I may also tweak the collision, right now its the triangle of the ship, but I worry that the speed of some stuff is too fast and it just over the triangle making up the fighter.

Really happy with the menuing I have to be totally honest. Dont plan to change that, and if I do more old games in the furture, I think I am just gonna steal this and just tweak it slightly.

TODOS:
1. Add Challenge stages
2. Add 'hatching' enemies
3. Look at fighter collision rework
4. Find a way to scale up difficulty